Tue, Oct 4, 2022, 1:00 PM +00:00

+++ pset4/filter/helpers.c
@@ -10,7 +10,7 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
         for (int j = 0; j <= width; j++)
         {
 
-            int soma    = image[i][j].rgbtGreen + image[i][j].rgbtRed + image[i][j].rgbtBlue;
+            float soma    = image[i][j].rgbtGreen + image[i][j].rgbtRed + image[i][j].rgbtBlue;
             int average = (int)round(soma / 3);
 
             image[i][j].rgbtGreen   = average;
